refactor: MyStudyPage에서 useMyStudyPage 훅 분리
