+++ frontend/src/pages/my-study-page/MyStudyPage.tsx
@@ -1,37 +1,11 @@
-import { useMemo } from 'react';
-
 import { css } from '@emotion/react';
-
-import type { MyStudy, StudyStatus } from '@custom-types';
 
 import Divider from '@components/divider/Divider';
 import Wrapper from '@components/wrapper/Wrapper';
 
 import * as S from '@my-study-page/MyStudyPage.style';
 import MyStudyCardListSection from '@my-study-page/components/my-study-card-list-section/MyStudyCardListSection';
-import useGetMyStudy from '@my-study-page/hooks/useGetMyStudy';
-
-const filterStudiesByStatus = (studies: Array<MyStudy>, status: StudyStatus) => {
-  return studies.filter(({ studyStatus }) => studyStatus === status);
-};
-
-const useMyStudyPage = () => {
-  const myStudyQueryResult = useGetMyStudy();
-
-  const filteredStudies: Record<string, Array<MyStudy>> = useMemo(() => {
-    const studies = myStudyQueryResult.data?.studies ?? [];
-    return {
-      prepare: filterStudiesByStatus(studies, 'PREPARE'),
-      inProgress: filterStudiesByStatus(studies, 'IN_PROGRESS'),
-      done: filterStudiesByStatus(studies, 'DONE'),
-    };
-  }, [myStudyQueryResult.data]);
-
-  return {
-    myStudyQueryResult,
-    studies: filteredStudies,
-  };
-};
+import { useMyStudyPage } from '@my-study-page/hooks/useMyStudyPage';
 
 const MyStudyPage: React.FC = () => {
   const { myStudyQueryResult, studies } = useMyStudyPage();
